Use end of symbol, not start, for LSP queries
This is generally better since the end tends to be at least as specific
as the start. For example, with `self.data`, we want to perform queries
on the `data` part, not the `self` part. This can be accomplished by
using the column corresponding to the end of the symbol.