Rewrite perf parsing code in Rust
Still need to actually switch over the Python code to use it. The old
Python code is too slow given how large the perf data gets (~100 MB for
a short run). Also found some potential bugs or inconsistencies in the
parsing code as I was rewriting it. Should go back and fix those.